Split rust_connection into two files
This moves the ConnectionInner struct into its own file.
